Refactor DxfValuePair and enhance evaluation expressions
- Move DxfValuePair to its own file with XML docs and namespace.
- Add EvaluatedValue property to EvaluationExpression.
- Reorder and document Connection property in BlockGripLocationComponent.
- Remove unused using directive from EvaluationExpression.cs.
