+++ src/ACadSharp/Objects/Evaluations/EvaluationExpression.cs
@@ -1,5 +1,4 @@
 ﻿using ACadSharp.Attributes;
-using static ACadSharp.Objects.Evaluations.EvaluationGraph;
 
 namespace ACadSharp.Objects.Evaluations;
 
@@ -9,6 +8,9 @@ namespace ACadSharp.Objects.Evaluations;
 [DxfSubClass(DxfSubclassMarker.EvalGraphExpr)]
 public abstract class EvaluationExpression : CadObject
 {
+	/// <summary>
+	/// Gets or sets the evaluated value of the expression.
+	/// </summary>
 	public DxfValuePair EvaluatedValue { get; set; }
 
 	/// <summary>
@@ -28,19 +30,4 @@ public abstract class EvaluationExpression : CadObject
 
 	[DxfCodeValue(99)]
 	public int Value99 { get; set; }
-}
-
-public class DxfValuePair
-{
-	public DxfCode Code { get; }
-
-	public GroupCodeValueType GroupCodeValueType { get { return GroupCodeValue.TransformValue((int)Code); } }
-
-	public object Value { get; }
-
-	public DxfValuePair(DxfCode code, object value)
-	{
-		Code = code;
-		Value = value;
-	}
 }
